acc/dcl arc_move and .py files moved from src to scripts
